Move all the optical constraints into the one module.
