+++ fresnel.cabal
@@ -47,12 +47,8 @@ library
     Fresnel.Optic
     Fresnel.Prism
     Fresnel.Profunctor.Coexp
-    Fresnel.Profunctor.Getting
-    Fresnel.Profunctor.Lensing
-    Fresnel.Profunctor.Isoing
     Fresnel.Profunctor.Optical
     Fresnel.Profunctor.Recall
-    Fresnel.Profunctor.Reviewing
     Fresnel.Review
     Fresnel.Setter
     Fresnel.Traversal
