+++ src/Fresnel/Getter.hs
@@ -16,7 +16,7 @@ import Data.Profunctor
 import Data.Profunctor.Unsafe ((#.), (.#))
 import Fresnel.Bifunctor.Contravariant
 import Fresnel.Optic
-import Fresnel.Profunctor.Getting
+import Fresnel.Profunctor.Optical
 
 -- Getters
